lulc_rf_project/requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
rasterio>=1.3.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.62.0
pillow>=9.0.0
scipy>=1.7.0
